requirements.txt
opencv-python
easyocr
numpy
pillow
torch
torchvision
face-recognition
pymongo
mediapipe
pytesseract